backend/matratum/requirements.txt
Django==2.2.12
django-cleanup==4.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.8.6